Adds initial Sphinx configuration.
Sphinx [1] is "a tool that makes it easy to create intelligent and beautiful
documentation." It also has great support for internationalization [2].

Sphinx is a common tool for writing about and documenting Python code, even
used by the Python language itself.

The TL;DR: of using Sphinx is: to make HTML documents, run `make html`. To
make PDF documents, run `make latexpdf`. It's as easy as that!

[1]: https://www.sphinx-doc.org/en/master/
[2]: https://www.sphinx-doc.org/en/master/usage/advanced/intl.html