Localizes handouts into Spanish, French, and German.
With this commit, the power of Sphinx's gettext integration begins to shine.

Translators (or in this case, just me) populate `.po` files autogenerated by
`sphinx-intl`, and then Sphinx can build documentation in any of those
languages.

To extract translatable messages into `pot` files, use:

    $ make gettext

These `pot` files are just generated by Sphinx and are kept out of version
control.

To generate `po` files from the `pot` files created by the previous step, use:

    $ sphinx-intl update -p _build/gettext -l es -l de -l fr

Any valid language codes can be used in that command.

Later, when the content of the translatable strings change, you'll have to run:

    $ sphinx-intl update -p _build/gettext

This will update the `po` files with the new strings for translators.

Finally, to make the documentation in a specific language (say, Spanish), use:

    $ make -e SPHINXOPTS="-D language='es'" html
    $ make -e SPHINXOPTS="-D language='es'" latexpdf

That's somewhat hard to remember, so on its way is a helpful script.