Adds handout autogeneration script and documentation.
The `makedoc.sh` script will make HTML and PDF versions of the handouts in all
languages it knows about and store the output in `handouts-html` and
`handouts-pdf`. It's a pretty janky script though, so manually check the
output before committing.

The README contains instructions for handling translations and building to
multiple languages.